+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+-   [ENH] Add `to_fasta` method to `janitor.biology` for exporting a DataFrame of sequences to a FASTA file. - Issue #646 @naseem173
 - [DOC] Add documentation explaining when cumulative-event aggregation is preferable to range join aggregations. - Issue #1702 @sumangouda
 -   [ENH] Avoid materializing all unequal pairs in `conditional_join` with
     `keep="first"` or `keep="last"`. - Issue #1651, PR #1681 @tunglambk